+++ ta4j-core/src/main/java/org/ta4j/core/BaseTradingRecord.java
@@ -27,8 +27,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.ta4j.core.Trade.TradeType;
-import org.ta4j.core.cost.CostModel;
-import org.ta4j.core.cost.ZeroCostModel;
+import org.ta4j.core.analysis.cost.CostModel;
+import org.ta4j.core.analysis.cost.ZeroCostModel;
 import org.ta4j.core.num.Num;
 
 /**