+++ ta4j-core/src/main/java/org/ta4j/core/criteria/AbstractAnalysisCriterion.java
@@ -1,0 +1,42 @@
+/**
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2014-2017 Marc de Verdelhan, 2017-2021 Ta4j Organization & respective
+ * authors (see AUTHORS)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package org.ta4j.core.criteria;
+
+import org.ta4j.core.AnalysisCriterion;
+
+/**
+ * An abstract analysis criterion.
+ */
+public abstract class AbstractAnalysisCriterion implements AnalysisCriterion {
+
+    @Override
+    public String toString() {
+        String[] tokens = getClass().getSimpleName().split("(?=\\p{Lu})", -1);
+        StringBuilder sb = new StringBuilder();
+        for (int i = 0; i < tokens.length - 1; i++) {
+            sb.append(tokens[i]).append(' ');
+        }
+        return sb.toString().trim();
+    }
+}