+++ ta4j-core/src/test/java/org/ta4j/core/criteria/AbstractCriterionTest.java
@@ -1,0 +1,75 @@
+/**
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2014-2017 Marc de Verdelhan, 2017-2021 Ta4j Organization & respective
+ * authors (see AUTHORS)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package org.ta4j.core.criteria;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.function.Function;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.ta4j.core.AnalysisCriterion;
+import org.ta4j.core.CriterionFactory;
+import org.ta4j.core.num.DecimalNum;
+import org.ta4j.core.num.DoubleNum;
+import org.ta4j.core.num.Num;
+
+@RunWith(Parameterized.class)
+public abstract class AbstractCriterionTest {
+
+    protected final Function<Number, Num> numFunction;
+    protected final OpenedPositionUtils openedPositionUtils = new OpenedPositionUtils();
+    private final CriterionFactory factory;
+
+    /**
+     * Constructor.
+     *
+     * @param factory CriterionFactory for building an AnalysisCriterion given
+     *                parameters
+     */
+    public AbstractCriterionTest(CriterionFactory factory, Function<Number, Num> numFunction) {
+        this.factory = factory;
+        this.numFunction = numFunction;
+    }
+
+    @Parameterized.Parameters(name = "Test Case: {index} (0=DoubleNum, 1=DecimalNum)")
+    public static List<Function<Number, Num>> function() {
+        return Arrays.asList(DoubleNum::valueOf, DecimalNum::valueOf);
+    }
+
+    /**
+     * Generates an AnalysisCriterion given criterion parameters.
+     *
+     * @param params criterion parameters
+     * @return AnalysisCriterion given parameters
+     */
+    public AnalysisCriterion getCriterion(Object... params) {
+        return factory.getCriterion(params);
+    }
+
+    public Num numOf(Number n) {
+        return numFunction.apply(n);
+    }
+
+}