+++ ta4j-core/src/test/java/org/ta4j/core/criteria/ReturnOverMaxDrawdownCriterionTest.java
@@ -1,0 +1,119 @@
+/**
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2014-2017 Marc de Verdelhan, 2017-2021 Ta4j Organization & respective
+ * authors (see AUTHORS)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package org.ta4j.core.criteria;
+
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+import static org.ta4j.core.TestUtils.assertNumEquals;
+
+import java.util.function.Function;
+
+import org.junit.Before;
+import org.junit.Test;
+import org.ta4j.core.AnalysisCriterion;
+import org.ta4j.core.BaseTradingRecord;
+import org.ta4j.core.Position;
+import org.ta4j.core.Trade;
+import org.ta4j.core.TradingRecord;
+import org.ta4j.core.mocks.MockBarSeries;
+import org.ta4j.core.num.NaN;
+import org.ta4j.core.num.Num;
+
+public class ReturnOverMaxDrawdownCriterionTest extends AbstractCriterionTest {
+
+    private AnalysisCriterion rrc;
+
+    public ReturnOverMaxDrawdownCriterionTest(Function<Number, Num> numFunction) {
+        super((params) -> new ReturnOverMaxDrawdownCriterion(), numFunction);
+    }
+
+    @Before
+    public void setUp() {
+        this.rrc = getCriterion();
+    }
+
+    @Test
+    public void rewardRiskRatioCriterion() {
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 105, 95, 100, 90, 95, 80, 120);
+        TradingRecord tradingRecord = new BaseTradingRecord(Trade.buyAt(0, series), Trade.sellAt(1, series),
+                Trade.buyAt(2, series), Trade.sellAt(4, series), Trade.buyAt(5, series), Trade.sellAt(7, series));
+
+        double totalProfit = (105d / 100) * (90d / 95d) * (120d / 95);
+        double peak = (105d / 100) * (100d / 95);
+        double low = (105d / 100) * (90d / 95) * (80d / 95);
+
+        assertNumEquals(totalProfit / ((peak - low) / peak), rrc.calculate(series, tradingRecord));
+    }
+
+    @Test
+    public void rewardRiskRatioCriterionOnlyWithGain() {
+        MockBarSeries series = new MockBarSeries(numFunction, 1, 2, 3, 6, 8, 20, 3);
+        TradingRecord tradingRecord = new BaseTradingRecord(Trade.buyAt(0, series), Trade.sellAt(1, series),
+                Trade.buyAt(2, series), Trade.sellAt(5, series));
+        assertTrue(rrc.calculate(series, tradingRecord).isNaN());
+    }
+
+    @Test
+    public void rewardRiskRatioCriterionWithNoPositions() {
+        MockBarSeries series = new MockBarSeries(numFunction, 1, 2, 3, 6, 8, 20, 3);
+        assertTrue(rrc.calculate(series, new BaseTradingRecord()).isNaN());
+    }
+
+    @Test
+    public void withOnePosition() {
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 95, 95, 100, 90, 95, 80, 120);
+        Position position = new Position(Trade.buyAt(0, series), Trade.sellAt(1, series));
+
+        AnalysisCriterion ratioCriterion = getCriterion();
+        assertNumEquals((95d / 100) / ((1d - 0.95d)), ratioCriterion.calculate(series, position));
+    }
+
+    @Test
+    public void betterThan() {
+        AnalysisCriterion criterion = getCriterion();
+        assertTrue(criterion.betterThan(numOf(3.5), numOf(2.2)));
+        assertFalse(criterion.betterThan(numOf(1.5), numOf(2.7)));
+    }
+
+    @Test
+    public void testNoDrawDownForTradingRecord() {
+        final MockBarSeries series = new MockBarSeries(numFunction, 100, 105, 95, 100, 90, 95, 80, 120);
+        final TradingRecord tradingRecord = new BaseTradingRecord(Trade.buyAt(0, series), Trade.sellAt(1, series),
+                Trade.buyAt(2, series), Trade.sellAt(3, series));
+
+        final Num result = rrc.calculate(series, tradingRecord);
+
+        assertNumEquals(NaN.NaN, result);
+    }
+
+    @Test
+    public void testNoDrawDownForPosition() {
+        final MockBarSeries series = new MockBarSeries(numFunction, 100, 105, 95, 100, 90, 95, 80, 120);
+        final Position position = new Position(Trade.buyAt(0, series), Trade.sellAt(1, series));
+
+        final Num result = rrc.calculate(series, position);
+
+        assertNumEquals(NaN.NaN, result);
+    }
+}